getyuhrates_web/pyproject.toml
[project]
name = "getyuhrates-web"
version = "0.1.0"
description = "FastAPI web application for GetYuhRates currency conversion service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "jinja2>=3.1.4",
    "python-multipart>=0.0.12",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "basedpyright>=1.18.0",
    "pytest>=8.3.3",
    "httpx>=0.27.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.basedpyright]
typeCheckingMode = "standard"
pythonVersion = "3.12"
